Added "Dynamic camera mode"
Dynamic camera mode sets the camera to be looking at and positioned over the players center point.
Static camera mode can be achieved as of now with a toggle in the core.World constructor call in GameView.

The dynamic camera for now has a fixed height but zoom will be added.